Rename MenuManagerContext to MenuDataContext
