+++ admin/src/contexts/menu-data.js
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+const MenuDataContext = createContext();
+
+export default MenuDataContext;
