--- admin/src/contexts/menu-manager.js
@@ -1,5 +1,0 @@
-import { createContext } from 'react';
-
-const MenuManagerContext = createContext();
-
-export default MenuManagerContext;
