+++ admin/src/hooks/index.js
@@ -1,2 +1,2 @@
-export { default as useMenuManager } from './use-menu-manager';
+export { default as useMenuData } from './use-menu-data';
 export { default as usePluginConfig } from './use-plugin-config';
