+++ admin/src/hooks/use-menu-data.js
@@ -1,0 +1,7 @@
+import { useContext } from 'react';
+
+import { MenuDataContext } from '../contexts';
+
+const useMenuData = () => useContext( MenuDataContext );
+
+export default useMenuData;
