--- admin/src/hooks/use-menu-manager.js
@@ -1,7 +1,0 @@
-import { useContext } from 'react';
-
-import { MenuManagerContext } from '../contexts';
-
-const useMenuManager = () => useContext( MenuManagerContext );
-
-export default useMenuManager;
